add search, hide players not in squad, customization

+++ src/lib.rs
@@ -1,5 +1,5 @@
-use std::{error::Error, collections::HashMap, sync::{Mutex, MutexGuard}, io::Write, fs::File};
-use arcdps::{Agent, CombatEvent, StateChange, extras::{UserInfoIter, UserInfo, ExtrasAddonInfo}, callbacks::{ImguiCallback, OptionsWindowsCallback}, imgui::{Ui, TableColumnSetup, Id, TableColumnFlags}};
+use std::{collections::HashMap, sync::{Mutex, MutexGuard}, io::Write, fs::File, ops::DerefMut};
+use arcdps::{extras::{UserInfoIter, ExtrasAddonInfo}, callbacks::{ImguiCallback, OptionsWindowsCallback}, imgui::{Ui, TableColumnSetup, ColorEdit}};
 use once_cell::sync::Lazy;
 use toml::{map::Map, Value};
 
@@ -12,12 +12,15 @@ arcdps::export! {
     imgui: draw_window,
     extras_squad_update: squad_update,
     options_windows: options,
-    // options_end, // This gives our own tab in the extension options
+    options_end: options_tab,
 }
 
 struct Player {
     name: String,
-    comment: String
+    lowercase_name: String,
+    comment: String,
+    lowercase_comment: String,
+    in_squad: bool
 }
 
 impl Player {
@@ -56,14 +59,22 @@ impl PlayerVecMap {
         delete
     }
 
-    fn delete_user(&mut self, username: &str) {
-        let index = self.name_dict.remove(username).unwrap();
-        self.delete_at(index)
+    fn user_left(&mut self, username: &str) {
+        let delete = self.is_deletable(username);
+        if delete {
+            let index = self.name_dict.remove(username).unwrap();
+            self.delete_at(index)
+        }
+
+        if let Some(index) = self.name_dict.get(username) {
+            let player = &mut self.player_list[*index];
+            player.in_squad = false;
+        }
     }
-    
+
     fn delete_at(&mut self, index: usize) {
         self.player_list.remove(index);
-        
+
         // After deleting the elements in the vec, all elements after it are shifted to the left. Update the indices
         for (_, idx) in self.name_dict.iter_mut() {
             if *idx > index {
@@ -78,7 +89,8 @@ impl PlayerVecMap {
 
         // The indices will be in reverse order so we can delete
         // them in same order without shifting any to-delete elements
-        for player in self.player_list.iter().rev() {
+        for player in self.player_list.iter_mut().rev() {
+            player.in_squad = false;
             if player.comment == "" {
                 if let Some(idx) = self.name_dict.remove(&player.name) {
                     delete_list.push(idx)
@@ -91,21 +103,64 @@ impl PlayerVecMap {
         }
     }
 
-    fn add(&mut self, username: &str) {
-        let new_item_index = self.player_list.len();
-        self.name_dict.insert(username.to_string(), new_item_index);
-        self.player_list.push(Player {
-            name: username.to_string(),
-            comment: "".to_string()
-        });
+    fn join(&mut self, username: &str) {
+        let add = !self.name_dict.contains_key(username);
+        if add {
+            let new_item_index = self.player_list.len();
+            self.name_dict.insert(username.to_string(), new_item_index);
+            self.player_list.push(Player {
+                name: username.to_string(),
+                lowercase_name: username.to_lowercase(),
+                comment: "".to_string(),
+                lowercase_comment: "".to_string(),
+                in_squad: false
+            });
+        }
+
+        if let Some(index) = self.name_dict.get(username) {
+            let player = &mut self.player_list[*index];
+            player.in_squad = true;
+        };
+    }
+}
+
+struct Filters {
+    user_filter_str: String,
+    comment_filter_str: String
+}
+
+impl Filters {
+    fn new() -> Filters {
+        Filters {
+            user_filter_str: String::new(),
+            comment_filter_str: String::new()
+        }
+    }
+}
+
+struct Flags {
+    extras_initialized: bool,
+    display_window: bool,
+    show_all: bool,
+}
+
+impl Flags {
+    fn new() -> Flags {
+        Flags {
+            extras_initialized: false,
+            display_window: false,
+            show_all: false
+        }
     }
 }
 
 struct State {
     players: PlayerVecMap,
     self_name: String,
-    extras_initialized: bool,
-    display_window: bool,
+    flags: Flags,
+    filters: Filters,
+    inactive_color: [f32;4],
+    comment_size: [f32;2]
 }
 
 impl State {
@@ -113,8 +168,10 @@ impl State {
         State {
             players: PlayerVecMap::new(),
             self_name: "".to_string(),
-            extras_initialized: false,
-            display_window: false
+            flags: Flags::new(),
+            filters: Filters::new(),
+            inactive_color: DEFAULT_INACTIVE_COLOR,
+            comment_size: DEFAULT_COMMENT_SIZE
         }
     }
 }
@@ -125,6 +182,11 @@ const TMP_PATH: &'static str = "addons/arcdps/player_list.tmp";
 
 const PLAYERS: &'static str = "Players";
 const OPENED_WINDOW: &'static str = "WindowOpen";
+const INACTIVE_COLOR: &'static str = "InactiveColor";
+const SHOW_ALL: &'static str = "ShowAll";
+const COMMENT_SIZE: &'static str = "CommentSize";
+const DEFAULT_INACTIVE_COLOR: [f32;4] = [0.5,0.5,0.5,1.0];
+const DEFAULT_COMMENT_SIZE: [f32;2] = [300.0, 20.0];
 
 fn init() -> Result<(), String> {
     // May return an error to indicate load failure
@@ -141,10 +203,51 @@ fn init() -> Result<(), String> {
         Some(Value::Boolean(b)) => b,
         _ => false,
     };
+    let inactive_color = match config.remove(INACTIVE_COLOR) {
+        Some(Value::Array(mut arr)) => {
+            if arr.len() == 4 {
+                let a = arr.remove(3);
+                let b = arr.remove(2);
+                let g = arr.remove(1);
+                let r = arr.remove(0);
+                if let (Value::Float(r), Value::Float(g), Value::Float(b), Value::Float(a)) = (r,g,b,a) {
+                    [r as f32,g as f32,b as f32,a as f32]
+                } else {
+                    DEFAULT_INACTIVE_COLOR
+                }
+            } else {
+                DEFAULT_INACTIVE_COLOR
+            }
+        },
+        _ => DEFAULT_INACTIVE_COLOR,
+    };
+    let comment_size = match config.remove(COMMENT_SIZE) {
+        Some(Value::Array(mut arr)) => {
+            if arr.len() == 2 {
+                let h = arr.remove(1);
+                let w = arr.remove(0);
+                if let (Value::Float(w), Value::Float(h)) = (w, h) {
+                    [w as f32,h as f32]
+                } else {
+                    DEFAULT_COMMENT_SIZE
+                }
+            } else {
+                DEFAULT_COMMENT_SIZE
+            }
+        },
+        _ => DEFAULT_COMMENT_SIZE,
+    };
+    let show_all = match config.remove(SHOW_ALL) {
+        Some(Value::Boolean(b)) => b,
+        _ => false,
+    };
 
     let mut state = get_state();
     state.players = player_list;
-    state.display_window = display_window;
+    state.flags.display_window = display_window;
+    state.flags.show_all = show_all;
+    state.inactive_color = inactive_color;
+    state.comment_size = comment_size;
 
     Ok(())
 }
@@ -153,7 +256,7 @@ fn init_extras(_: ExtrasAddonInfo, self_name: Option<&str>) {
     let mut state = get_state();
 
     if let Some(self_name) = self_name {
-        state.extras_initialized = true;
+        state.flags.extras_initialized = true;
         state.self_name = self_name.to_owned();
     }
 }
@@ -180,8 +283,11 @@ fn init_player_list(config: &mut Map<String, Value>) -> PlayerVecMap {
 
             if let (Some(Value::String(name)), Some(Value::String(comment))) = (name, comment) {
                 Some(Player {
+                    lowercase_name: name.to_lowercase(),
                     name,
-                    comment
+                    lowercase_comment: comment.to_lowercase(),
+                    comment,
+                    in_squad: false,
                 })
             } else {
                 None
@@ -210,7 +316,14 @@ fn release() {
         }
     }).collect();
     config.insert(PLAYERS.to_string(), Value::Array(player_list));
-    config.insert(OPENED_WINDOW.to_string(), Value::Boolean(state.display_window));
+    config.insert(OPENED_WINDOW.to_string(), Value::Boolean(state.flags.display_window));
+    let inactive_color = state.inactive_color.into_iter()
+        .map(|val| Value::Float(val as f64)).collect();
+    config.insert(INACTIVE_COLOR.to_string(), Value::Array(inactive_color));
+    let comment_size = state.comment_size.into_iter()
+        .map(|val| Value::Float(val as f64)).collect();
+    config.insert(COMMENT_SIZE.to_string(), Value::Array(comment_size));
+    config.insert(SHOW_ALL.to_string(), Value::Boolean(state.flags.show_all));
 
     let toml_string = toml::to_string(&Value::Table(config)).unwrap();
     std::fs::write(CONFIG_PATH, toml_string).unwrap()
@@ -235,12 +348,11 @@ fn remove_user(username: &str) {
     let mut state = get_state();
 
     let is_self = username == state.self_name;
-    let delete = state.players.is_deletable(username);
 
     if is_self {
         state.players.delete_all()
-    } else if delete {
-        state.players.delete_user(username);
+    } else {
+        state.players.user_left(username);
     }
 }
 
@@ -248,11 +360,9 @@ fn add_user(username: &str) {
     let mut state = get_state();
 
     let is_self = username == state.self_name;
-    let add = !state.players.name_dict.contains_key(username);
 
-    // Only add if it's not there
-    if !is_self && add {
-        state.players.add(username);
+    if !is_self {
+        state.players.join(username);
     }
 }
 
@@ -263,7 +373,7 @@ fn draw_window(ui: &Ui, not_character_or_loading: bool) {
         return
     }
 
-    if !state.extras_initialized {
+    if !state.flags.extras_initialized {
         arcdps::imgui::Window::new("Player List Error").build(ui, || {
             ui.text("Unofficial extras extension required")
         });
@@ -271,10 +381,9 @@ fn draw_window(ui: &Ui, not_character_or_loading: bool) {
         return
     };
 
-    let mut opened_window = state.display_window;
+    let mut opened_window = state.flags.display_window;
     std::mem::drop(state); // liberates the mutex so get_state() can be called again from the closure in .build()
     if opened_window {
-
         arcdps::imgui::Window::new("Player List").opened(&mut opened_window).collapsible(false).build(ui, || {
             let column_data = [
                 // max character length of account name = 32 characters
@@ -287,30 +396,75 @@ fn draw_window(ui: &Ui, not_character_or_loading: bool) {
                     ..Default::default()
                 }
             ];
+            {
+                let mut state = get_state();
+                ui.checkbox("Show all", &mut state.flags.show_all);
+                ui.text("Filters:");
+                if ui.input_text("##user_filter", &mut state.filters.user_filter_str).build() {
+                    state.filters.user_filter_str = state.filters.user_filter_str.to_lowercase()
+                };
+                if ui.is_item_hovered() {
+                    ui.tooltip(|| ui.text("Filter by user name"))
+                }
+                if ui.input_text("##comment_filter", &mut state.filters.comment_filter_str).build() {
+                    state.filters.comment_filter_str = state.filters.comment_filter_str.to_lowercase()
+                };
+                if ui.is_item_hovered() {
+                    ui.tooltip(|| ui.text("Filter by comment"))
+                }
+            }
             if let Some(table) = ui.begin_table_header("PLayerListTable", column_data) {
                 let mut state = get_state();
-                for (i, player) in state.players.player_list.iter_mut().enumerate() {
+                let state = state.deref_mut();
+                let filters = &state.filters;
+                let players = &mut state.players;
+                for (i, player) in players.player_list.iter_mut().enumerate() {
+                    if !filters.user_filter_str.is_empty() && !player.lowercase_name.starts_with(&filters.user_filter_str) {
+                        continue;
+                    }
+                    if !filters.comment_filter_str.is_empty() && !player.lowercase_comment.starts_with(&filters.comment_filter_str) {
+                        continue;
+                    }
+                    if !state.flags.show_all && !player.in_squad {
+                        continue;
+                    }
                     ui.table_next_column();
-                    ui.text(&player.name);
+                    if player.in_squad {
+                        ui.text(&player.name);
+                    } else {
+                        ui.text_colored(state.inactive_color, &player.name)
+                    }
                     ui.table_next_column();
-                    ui.input_text_multiline(format!("##{i}"), &mut player.comment, [80.0, 40.0]).build();
+                    if ui.input_text_multiline(format!("##{i}"), &mut player.comment, state.comment_size).build() {
+                        player.lowercase_comment = player.comment.to_lowercase()
+                    };
                 }
                 table.end()
             };
         });
     }
 
-    get_state().display_window = opened_window;
+    get_state().flags.display_window = opened_window;
 }
 
 fn options(ui: &Ui, window_name: Option<&str>) -> bool {
     if let Some("error") = window_name {
-        ui.checkbox("player list", &mut get_state().display_window);
+        ui.checkbox("player list", &mut get_state().flags.display_window);
     }
 
     false
 }
 
+fn options_tab(ui: &Ui) {
+    let mut state = get_state();
+    ColorEdit::new("Inactive player", &mut state.inactive_color).build(ui);
+    if ui.is_item_hovered() {
+        ui.tooltip(|| ui.text("Color of the names of players out of the squad"))
+    }
+
+    ui.input_float2("Comment Size", &mut state.comment_size).build();
+}
+
 fn log(msg: &str) {
-    writeln!(File::options().append(true).open(TMP_PATH).unwrap(), "{msg}").unwrap();
+    writeln!(File::options().create(true).append(true).open(TMP_PATH).unwrap(), "{msg}").unwrap();
 }
